fix(vbe): change framebuffer_addr type from uint64_t to uint32_t for consistency and optrimisation of vbe_draw_*

+++ drivers/display/vbe/vbe.h
@@ -13,7 +13,7 @@ enum vbe_result {
 };
 
 typedef struct s_vbe_info {
-    uint64_t framebuffer_addr;
+    uint32_t framebuffer_addr;
     uint32_t width;
     uint32_t height;
     uint32_t pitch;
